Cleans utils file
