+++ assets/js/helpers/utils.js
@@ -26,4 +26,11 @@ const keyup = (key, callback) => {
   });
 };
 
-export { keyup }; // eslint-disable-line
+const preloadImage = url => new Promise((resolve, reject) => {
+  const image = new Image();
+  image.src = url;
+  image.onload = resolve;
+  image.onerror = reject;
+});
+
+export { keyup, preloadImage };
